microservices/jdemetra-io-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jdemetra-io-service"
version = "0.1.0"
description = "Data I/O service for JDemetra+"
readme = "README.md"
authors = [
    {name = "JDemetra+ Team"},
]
license = {text = "EUPL-1.2"}
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "pydantic>=2.0",
    "numpy>=2.0",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "lxml>=4.9",
    "pyyaml>=6.0",
    "minio>=7.2",
    "aiofiles>=23.2",
    "jdemetra-common @ file:../jdemetra-common",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=2.0",
    "httpx>=0.25",
    "black>=22.0",
    "mypy>=0.950",
]